+++ Project/frontend/src/components/Tips/TipsItemDetail.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-container v-if="isTaken" fill-height fluid grid-list-md>
+  <v-container v-if="isTaken" fill-height fluid grid-list-md style="font-family: 'Handon3gyeopsal300g' !important;">
     <!-- {{ tip }} -->
     <!-- <v-layout v-if="loading" align-center justify-center>
       <v-progress-circular size="50" color="primary" indeterminate></v-progress-circular>
@@ -11,23 +11,8 @@
         <v-layout justify-start column fill-height>
           <v-flex xs12>
             <v-card>
-              <v-card-title class="headline pb-0">
+              <v-card-title class="headline pb-0" style="font-family: 'Handon3gyeopsal600g' !important;">
                 {{ tip.title }}
-                <v-spacer></v-spacer>
-                <v-toolbar v-if="$store.state.userInfo.id === tip.writer" flat color="white">
-                  <v-row justify="end">
-                    <template>
-                      <v-btn @click="goTipsUpdate" icon>
-                        <v-icon>mdi-pencil</v-icon>
-                      </v-btn>
-                      <v-btn icon @click="goTipsDelete">
-                        <v-icon>
-                          mdi-delete
-                        </v-icon>
-                      </v-btn>
-                    </template>
-                  </v-row>
-                </v-toolbar>
               </v-card-title>
               <v-layout>
                 <v-flex xs7 class='pr-0'>
@@ -35,56 +20,68 @@
                     <v-list-item>
                       <v-list-item-avatar>
                       <!-- <img :src="'https://steemitimages.com/u/' + author + '/avatar/small'" alt="avatar" onerror="this.src='https://steemitimages.com/u/monawoo/avatar/small'"> -->
-                      <img :src="require('../../../src/assets/Totodile.jpg')" alt="avatar">
+                        <v-icon v-if="psas[tip.writer]==null" x-large>mdi-account-circle</v-icon>
+                        <img v-else :src="psas[tip.writer]" alt="avatar">
                       </v-list-item-avatar>
                       <v-list-item-content>
-                        <v-list-item-title>{{users[tip.writer]}}
+                        <v-list-item-title style="font-family: 'Handon3gyeopsal600g';">{{users[tip.writer]}}
                           <!-- <span>({{ tip.writer }})</span> -->
                         </v-list-item-title>
-                        <v-list-item-title>{{tip.created_at | filterCreated }}</v-list-item-title>
+                        <v-list-item-title class="grey--text">{{tip.created_at | filterCreated }}</v-list-item-title>
                         <!-- <v-list-item-sub-title>{{tip.created_at | filterCreated}} · {{category}}</v-list-item-sub-title> -->
                       </v-list-item-content>
                     </v-list-item>
                   </v-list>
                 </v-flex>
-                <v-flex xs5 text-xs-right class='pr-4 pt-3'>
-                  <div v-if="!liked">
-                    <v-btn @click="like" icon><v-icon>mdi-heart-outline</v-icon></v-btn>{{ tip.likeAccounts.length }}명이 이 글을 좋아합니다.
-                  </div>
-                  <div v-else-if="liked && tip.likeAccounts.length == 1">
-                    <v-btn @click="like" icon><v-icon color="red darken-1">mdi-heart</v-icon></v-btn>{{ userInfo.nickname }}님 이 글을 좋아합니다.
-                  </div>
-                  <div v-else>
-                    <v-btn @click="like" icon><v-icon color="red darken-1">mdi-heart</v-icon></v-btn>{{ userInfo.nickname }}님 외 {{ tip.likeAccounts.length - 1 }} 명이 글을 좋아합니다.
-                  </div>
-                  <div>
-                    · 댓글 {{ tip.comments.length }}명
-                  </div>
+                <v-flex xs5 text-xs-right class='pt-3'>
+                  <v-btn v-if="!liked" @click="like" icon><v-icon>mdi-heart-outline</v-icon></v-btn>
+                  <v-btn v-else @click="like" icon><v-icon color="red darken-1">mdi-heart</v-icon></v-btn>{{ tip.likeAccounts.length }}
+                  <v-btn icon><v-icon>mdi-comment-processing-outline</v-icon></v-btn>{{ tip.comments.length }}
+                  <v-btn icon v-if="!scraped" @click="scrap"><v-icon>mdi-bookmark-outline</v-icon></v-btn>
+                  <v-btn icon v-else color="teal"><v-icon>mdi-bookmark</v-icon></v-btn>
                 </v-flex>
               </v-layout>
               <v-divider></v-divider>
               <v-card-text>
-                <article>{{ tip.contents }}</article>
+                <v-img v-if="tip.img_src!=null" class="mt-3" width="100%" height="auto" :src="tip.img_src"></v-img>
+                <article class="mt-5 black--text">{{ tip.contents }}</article>
               </v-card-text>
-              <v-card-text>
+              <v-card-text class="pb-1">
                 <template>
-                  <v-chip href='javascript:false' class='tag'>#{{ tag[tip.tag_id] }}</v-chip>
+                  <v-toolbar v-if="$store.state.userInfo.id === tip.writer" flat color="white">
+                    <v-row justify="center" align="center">
+                      <template>
+                      <v-chip class="tag">#{{ tag[tip.tag_id] }}</v-chip>
+                      </template>
+                      <v-spacer></v-spacer>
+                      <template>
+                        <v-btn @click="goTipsUpdate" icon>
+                          <v-icon>mdi-pencil</v-icon>
+                        </v-btn>
+                        <v-btn icon @click="goTipsDelete">
+                          <v-icon>
+                            mdi-delete
+                          </v-icon>
+                        </v-btn>
+                      </template>
+                    </v-row>
+                  </v-toolbar>
                 </template>
               </v-card-text>
             </v-card>
           </v-flex>
           <v-flex xs12>
             <!-- <v-subheader class='pl-0' >댓글 ({{tip.comments.length}})</v-subheader> -->
-            <v-subheader class='pl-0' >댓글 ({{tip.comments.length}})</v-subheader>
+            <v-subheader class='pl-0' >댓글</v-subheader>
             <v-card ref='comments'>
-              <v-card-text>
+              <v-card-text class="pb-0">
                 <v-text-field
                   ref="contents"
                   v-model="new_comment.contents"
-                  label="Content"
                   placeholder="댓글을 입력해 주세요."
                   required
                   @keyup.enter="newComment"
+                  color="green darken-2"
                 >
                 </v-text-field>
               </v-card-text>
@@ -117,6 +114,7 @@
 import axios from 'axios'
 import { mapState } from 'vuex'
 import TipComment from '../Tips/TipComment.vue'
+import firebase from 'firebase'
 
 export default {
   name: 'TipsItemDetail',
@@ -126,8 +124,8 @@ export default {
   data() {
     return {
       tip: {},
-      // comments: [],
       liked: false,
+      scraped: false,
       new_comment : {
         contents: '',
         writer: this.$store.state.userInfo.id,
@@ -137,18 +135,17 @@ export default {
       tag: {1:'요리',2:'세탁',3:'청소',4:'보관'}
     }
   },
-  // 아직 등록자와 작성자가 같은지 확인은 안함...
   methods: {
     goTipsDelete() {
       if (this.$store.state.userInfo.id === this.tip.writer) {
-        axios.delete(`${this.$store.state.api_server}/tips/${this.tip.id}`)
-          .then(() => {
-            // console.log(res.data)
-            this.$router.push('/tips')
-          })
-          .catch(err => {
-            console.log(err)
-          })
+        const response = confirm('게시글을 정말로 삭제하시겠습니까?')
+        if (response) {
+          axios.delete(`${this.$store.state.api_server}/tips/${this.tip.id}`)
+            .then(() => {
+              firebase.storage().ref().child(`tips/${this.tip.id}`).delete()
+              this.$router.push('/tips')
+            })
+        }
       } else {
         alert('게시글 작성자만 삭제 가능 합니다.')
       }
@@ -161,20 +158,21 @@ export default {
     like() {
       axios.post(`${this.$store.state.api_server}/tips/${this.tip.id}/like`, '',{ params : { user_id: this.$store.state.userInfo.id }})
         .then(() => {
-          // console.log(res)
           if (this.liked) {
             this.liked = false
-            // console.log(this.tip.likeAccounts)
             this.tip.likeAccounts.splice(this.tip.likeAccounts.indexOf(this.$store.state.userInfo.id),1)
-            // console.log(this.tip.likeAccounts)
           } else {
             this.liked = true
             this.tip.likeAccounts.push(this.$store.state.userInfo.id)
-            // console.log(this.tip.likeAccounts)
           }
         })
-        .catch(err => {
-          console.log(err.message)
+    },
+    scrap() {
+      axios.post(`${this.$store.state.api_server}/tips/${this.tip.id}/scrap`, '', { params: { user_id: this.$store.state.userInfo.id}})
+        .then(() => {
+          this.tip.scarpAccounts.push(this.$store.state.userInfo.id)
+          this.scraped = true
+          alert('팁이 저장되었습니다.')
         })
     },
     deleteComment(commentId) {
@@ -182,20 +180,12 @@ export default {
         .then(() => {
           this.tip.comments.splice(this.tip.comments.findIndex(x => x.id === commentId),1)
         })
-        .catch(err => {
-          console.log(err)
-        })
     },
     newComment() {
       axios.post(`${this.$store.state.api_server}/tips/${this.tip.id}/comments`, this.new_comment, { params: {tip_id: this.tip.id }})
         .then(res => {
-          // console.log(res.data)
           this.tip.comments.push(res.data)
-          // console.log(this.tip.comments)
           this.new_comment.contents = ''
-        })
-        .catch(err => {
-          console.log(err.message)
         })
     },
     updateComment(changeComment) {
@@ -203,31 +193,27 @@ export default {
         .then(res => {
           this.tip.comments.splice(this.tip.comments.findIndex(x => x.id === changeComment.id), 1, res.data)
         })
-        .catch(err => {
-          console.log(err)
-        })
     }
   },
   created() {
     axios.get(`${this.$store.state.api_server}/tips/${this.$route.params.tip_id}`)
       .then(res => {
-        // console.log(res.data)
         this.tip = res.data
         if (this.tip.likeAccounts.length>0 && this.tip.likeAccounts.includes(this.$store.state.userInfo.id)) {
           this.liked = true
         }
+        if (res.data.scrapAccounts.includes(this.$store.state.userInfo.id)) {
+          this.scraped = true
+        }
         this.isTaken = true
-      })
-      .catch(err => {
-        // console.log(this.id)
-        console.log(err)
       })
   },
   computed: {
     ...mapState([
       'api_server',
       'users',
-      'userInfo'
+      'userInfo',
+      'psas'
     ])
   }
 }
